requirements.txt
ttkbootstrap
pytesseract
Pillow
opencv-python
pandas
openpyxl
